Refactor into hard75challenge.py
